maybe maybe

+++ FUZZER_HW/src/main.c
@@ -1,15 +1,16 @@
 #include "../include/fuzzer.h"
 #include "../include/global.h"
+#include <stddef.h>
 #include <stdlib.h>
 #include <string.h>
 
 int main(int argc, char *argv[]) {
-    int jobs = DEFAULT_RUNNER_COUNT;
-    int inputs = DEFAULT_INPUT_TOTAL;
-    int tlimit = DEFAULT_TIMEOUT_SEC;
+    size_t jobs = DEFAULT_RUNNER_COUNT;
+    size_t inputs = DEFAULT_INPUT_TOTAL;
+    size_t tlimit = DEFAULT_TIMEOUT_SEC;
     char *seed_path = NULL;
 
-    int i = 1;
+    size_t i = 1;
     for (; i < argc && argv[i][0] == '-' && argv[i][1] != '\0'; i++) {
         char flag = argv[i][1];
         if (flag == 'h') {
